Add create favorite recipe functionality
