+++ spec/features/recipes/show_spec.rb
@@ -2,9 +2,10 @@ require 'rails_helper'
 
 RSpec.describe User do
   it 'can view a recipe' do
-    VCR.use_cassette("blackened_salmon_recipe") do
+    VCR.use_cassette("salmon_recipe") do
       user = create(:user)
       allow_any_instance_of(ApplicationController).to receive(:current_user).and_return(user)
+
       visit recipe_path(4584)
 
       expect(page).to have_content('Blackened Salmon With Hash Browns and Green Onions')
